+++ DPLocalization/Plural/Plural+DPLocalization.m
@@ -1,6 +1,6 @@
 //
 //  Plural+DPLocalization.m
-//  LocalizationDemo
+//  DP Commons
 //
 //  Created by Dmitriy Petrusevich on 25/06/15.
 //  Copyright (c) 2015 Dmitriy Petrusevich. All rights reserved.
@@ -52,19 +52,84 @@ dp_plural_rules_func dp_plural_rules_for_lang_code(NSString *lang_code) {
     return ptr ? ptr : dp_plural_rules_always_other;
 }
 
-#pragma mark -
+#pragma mark - NSNumber
 
 @implementation NSNumber (DPLocalization_Plural)
 
 - (enum DPPluralRule)pluralRuleWithLanguage:(NSString *)language {
-    int value = ABS([self intValue]);
     dp_plural_rules_func func = dp_plural_rules_for_lang_code(language);
-    return func(value, value, 0, 0, 0, 0);
+    return [self pluralRuleWithRules:func];
 }
 
 - (enum DPPluralRule)pluralRuleWithRules:(dp_plural_rules_func)rules {
-    int value = ABS([self intValue]);
-    return rules(value, value, 0, 0, 0, 0);
+    double n = ABS([self doubleValue]);
+    int i = (int)n;
+    return rules(n, i, 0, 0, 0, 0);
+}
+
+@end
+
+#pragma mark - DPFormattedValue
+
+@implementation DPFormattedValue (DPLocalization_Plural)
+
+- (enum DPPluralRule)pluralRuleWithLanguage:(NSString *)language {
+    dp_plural_rules_func func = dp_plural_rules_for_lang_code(language);
+    return [self pluralRuleWithRules:func];
+}
+
+- (enum DPPluralRule)pluralRuleWithRules:(dp_plural_rules_func)rules {
+    enum DPPluralRule result = DPPluralRuleUnknown;
+
+    if ([[self value] isKindOfClass:[NSNumber class]]) {
+        NSNumber *value = [self value];
+
+        if ([[self formatter] isKindOfClass:[NSNumberFormatter class]]) {
+            NSNumberFormatter *formatter = (NSNumberFormatter *)[self formatter];
+            NSString *strValue = [self description];
+            NSString *decimalSeparator = [formatter decimalSeparator];
+            
+            double n = ABS([value doubleValue]);
+            int i = (int)n;
+            int v = 0, w = 0, f = 0, t = 0;
+            
+            NSRange separatorRange = [strValue rangeOfString:decimalSeparator];
+            if (separatorRange.location != NSNotFound) {
+                NSUInteger startPosition = separatorRange.location + separatorRange.length;
+                NSUInteger endPosition = startPosition;
+                
+                for (; endPosition < strValue.length; endPosition++) {
+                    unichar ch = [strValue characterAtIndex:endPosition];
+                    if (ch < '0' || ch > '9') {
+                        break;
+                    }
+                }
+                
+                v = (int)(endPosition - startPosition);
+                NSString *fractionalPart = [strValue substringWithRange:NSMakeRange(startPosition, v)];
+                f = [fractionalPart intValue];
+                
+                NSInteger lastNonZeroCharacterIndex = (fractionalPart.length - 1);
+                for (; lastNonZeroCharacterIndex >= 0; lastNonZeroCharacterIndex--) {
+                    if ([fractionalPart characterAtIndex:lastNonZeroCharacterIndex] != '0') {
+                        break;
+                    }
+                }
+                
+                NSString *fractionalPartWithoutZeros = [fractionalPart substringToIndex:(lastNonZeroCharacterIndex + 1)];
+                w = (int)[fractionalPartWithoutZeros length];
+                t = [fractionalPartWithoutZeros intValue];
+            }
+            
+            
+            return rules(n, i, v, w, f, t);
+        }
+        else {
+            result = [value pluralRuleWithRules:rules];
+        }
+    }
+
+    return result;
 }
 
 @end